backend/requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
flask-cors
bcrypt
requests
python-dotenv
gunicorn